requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3

# Document processing
PyPDF2==3.0.1
python-docx==1.0.0

# Vector database
qdrant-client>=1.14.2
faiss-cpu==1.7.4  # Alternative to Qdrant

# Azure OpenAI
openai==0.28.1

# Frontend
streamlit==1.28.0

# Utilities
numpy>=1.26.1
pandas>=2.1.1
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
